mitäs kettua

+++ sql/create_tables.sql
@@ -29,10 +29,10 @@ CREATE TABLE viini (
   kotimaa_id     INT REFERENCES kotimaa (id),
   nimi           VARCHAR(255)                  NOT NULL,
   vuosikerta     SMALLINT CHECK (vuosikerta > 1800),
-  alkoholi       DECIMAL CHECK (alkoholi >= 0) NOT NULL,
-  happo          DECIMAL CHECK (happo >= 0)    NOT NULL,
+  alkoholi       REAL CHECK (alkoholi >= 0) NOT NULL,
+  happo          REAL CHECK (happo >= 0)    NOT NULL,
   makeus         SMALLINT CHECK (makeus >= 4), -- Makeus on joko 4g/l tai yli, muuten null
-  uutos          SMALLINT CHECK (uutos > 0)
+  uutos          REAL CHECK (uutos > 0)
 
 
 );
